+++ weblate/trans/models/component.py
@@ -11,7 +11,7 @@ from collections import defaultdict
 from contextlib import suppress
 from glob import glob
 from itertools import chain
-from typing import TYPE_CHECKING, Any, TypedDict, cast
+from typing import TYPE_CHECKING, Any, ClassVar, TypedDict, cast
 from urllib.parse import quote as urlquote
 from urllib.parse import urlparse
 
@@ -427,6 +427,15 @@ class OldComponentSettings(TypedDict):
 class Component(  # noqa: PLR0904
     models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin, LockMixin
 ):
+    LINKED_REPOSITORY_SETTINGS: ClassVar[tuple[str, ...]] = (
+        "push_on_commit",
+        "commit_pending_age",
+        "auto_lock_error",
+    )
+    LINKED_REPOSITORY_SETTING_MESSAGE = gettext_lazy(
+        "Option is not available for linked repositories. Setting from linked component will be used."
+    )
+
     name = models.CharField(
         verbose_name=gettext_lazy("Component name"),
         max_length=COMPONENT_NAME_LENGTH,
@@ -1603,6 +1612,27 @@ class Component(  # noqa: PLR0904
         """Return absolute shareable URL."""
         return self.project.get_share_url()
 
+    @property
+    def effective_repo_component(self) -> Component:
+        linked = self.linked_component
+        if linked is not None:
+            if linked.project_id == self.project_id:
+                linked.project = self.project
+            return linked
+        return self
+
+    @property
+    def effective_push_on_commit(self) -> bool:
+        return self.effective_repo_component.push_on_commit
+
+    @property
+    def effective_commit_pending_age(self) -> int:
+        return self.effective_repo_component.commit_pending_age
+
+    @property
+    def effective_auto_lock_error(self) -> bool:
+        return self.effective_repo_component.auto_lock_error
+
     @perform_on_link
     def _get_path(self):
         """Return full path to component VCS repository."""
@@ -2077,7 +2107,7 @@ class Component(  # noqa: PLR0904
         * Configured push
         * Whether there is something to push
         """
-        if not self.push_on_commit:
+        if not self.effective_push_on_commit:
             self.log_info("skipped push: push on commit disabled")
             return
         if not self.can_push():
@@ -3012,7 +3042,7 @@ class Component(  # noqa: PLR0904
 
     @property
     def lock_alerts(self):
-        if not self.auto_lock_error:
+        if not self.effective_auto_lock_error:
             return []
         return [
             alert for alert in self.all_active_alerts if alert.name in LOCKING_ALERTS
@@ -3035,7 +3065,7 @@ class Component(  # noqa: PLR0904
             self.clear_prefetched_alerts()
             if (
                 self.locked
-                and self.auto_lock_error
+                and self.effective_auto_lock_error
                 and alert in LOCKING_ALERTS
                 and not self.alert_set.filter(name__in=LOCKING_ALERTS).exists()
                 and getattr(
@@ -3064,7 +3094,7 @@ class Component(  # noqa: PLR0904
             self.all_alerts[alert] = obj
 
         # Automatically lock on error
-        if created and self.auto_lock_error and alert in LOCKING_ALERTS:
+        if created and self.effective_auto_lock_error and alert in LOCKING_ALERTS:
             self.do_lock(user=None, lock=True, auto=True)
 
         # Update details with exception of component removal
@@ -3555,12 +3585,7 @@ class Component(  # noqa: PLR0904
             for setting in ("push", "branch", "push_branch"):
                 if getattr(self, setting):
                     raise ValidationError(
-                        {
-                            setting: gettext(
-                                "Option is not available for linked repositories. "
-                                "Setting from linked component will be used."
-                            )
-                        }
+                        {setting: self.LINKED_REPOSITORY_SETTING_MESSAGE}
                     )
         # Make sure we are not using stale link even if link is not present
         self.linked_component = Component.objects.get_linked(self.repo)